Work in a no-std environment
Fixes #19
